feat: logout swagger 적용 및 폴더 구조 변경
- 회원 가입과 로그아웃 및 로그인 swagger 모아둠.
- logout swagger 적용
